+++ dd-java-agent/instrumentation/spring-webmvc-3.1/src/bootTest/groovy/test/ContainerType.groovy
@@ -1,0 +1,32 @@
+package test
+
+import org.springframework.boot.context.embedded.EmbeddedServletContainer
+import org.springframework.boot.context.embedded.jetty.JettyEmbeddedServletContainer
+import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainer
+import org.springframework.boot.context.embedded.undertow.UndertowEmbeddedServletContainer
+import test.boot.SpringBootBasedTest
+
+enum ContainerType {
+  TOMCAT("tomcat-server"),
+  JETTY("jetty-server"),
+  UNDERTOW("unknown"),
+  DEFAULT("java-web-servlet");
+
+  final String component
+
+    ContainerType(String component) {
+    this.component = component
+  }
+
+  static ContainerType forEmbeddedServletContainer(EmbeddedServletContainer container) {
+    if (container instanceof TomcatEmbeddedServletContainer) {
+      return TOMCAT
+    } else if (container instanceof JettyEmbeddedServletContainer) {
+      return JETTY
+    } else if (container instanceof UndertowEmbeddedServletContainer) {
+      return UNDERTOW
+    } else {
+      return DEFAULT
+    }
+  }
+}
