+++ dd-java-agent/instrumentation/spring-webmvc-3.1/src/bootTest/groovy/test/boot/AppConfig.groovy
@@ -1,11 +1,7 @@
 package test.boot
 
-import org.apache.catalina.connector.Connector
+
 import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.embedded.EmbeddedServletContainerFactory
-import org.springframework.boot.context.embedded.tomcat.TomcatConnectorCustomizer
-import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory
-import org.springframework.context.annotation.Bean
 import org.springframework.web.servlet.config.annotation.PathMatchConfigurer
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 import org.springframework.web.util.UrlPathHelper
